Changed the averageAnswerTime type to float in PlayerResults struct

+++ Trivia_Ofer_And_Shaked_Sisso_2023/Responses.h
@@ -108,7 +108,7 @@ typedef struct PlayerResults
 	string username;
 	unsigned int correctAnswerCount;
 	unsigned int wrongAnswerCount;
-	unsigned int averageAnswerTime;
+	float averageAnswerTime;
 } PlayerResults;
 
 typedef struct GetGameResultsResponse
